Remove pseudo-compatibility function shinken.utils.if_else().
This function was introduced in
2012-05-01T21:44:00Z!gerhard.lausser@consol.de as "a ternary operator
workaround ensures compatibility with python 2.4".

Using a function has a major drawback: all values passed as arguments have
to be evaluated prior to the call. In opposite to the ternity operator and
the well-known `... and ... or` pattern, where only the resulting part gets
evaluated.

It is used only in shinken.log and one of the values passed is
`time.asctime(time.localtime(time.time()))` which is quite heavy for being
evaluated without using it.

So I remove this function at all and replace it by either boring if-else-
statements or the `... and ... or` pattern.