According with the shinken owner, removing the libexec third part of the shinken sources.
Modification of the windows installation script to download the thirdpart from veosoft. Adding the sources of wget and wmic programs to match the agpl v3 license. adding the dounzip script to unzip the thirdpart without external program (no command line on windows to do that). Another time compliance with the agpl to use only other agpl programs into the source folder (or only use other programs, but in the windows world, the others tools are not given or in a license mode not directly compatible ith the agpl one) and display the sources. My main goal is to use a realy out of the box usable script....