For the CMD files, Put the python exec file as a var. If the .py is recognized on windows, not setting the var will launch the .py ;-) It is also a good thing to try multiple versions for python exec....
Fix a bug to unzip files in the dounzip.vbs script
Change the location of ZIP files on the veosoft server for the 1.4 release...
Insert pause in the script to check the process during the RC release.
Adding Check_winping.exe programm and changoing the command for a windows 2k8 host template from windows (check_tcp is not working on windows)
Preparing new script for a better management of installations and uninstallation from script.
Some new fix on vars and locations....