Added service instance properties override feature
It allows to override specific service instances properties through the
definition of the new `serviceoverride` object. This is mainly useful for
dynamycally generated services such as those linked to host templates (those
defined in packs for instances), or linked to hostgroups. This allows to
override a specific instance parameters without having to define a complete
new service.