Added the ability to hide service in problems view
If both an host/service instance has the `hide_in_ui_problems` flag set and a
user has the `hide_ui_problems` flag set, the host/service is hidden to the
user in the webui problems view, even if he's set as contact to this
host/service.

This allows to hide services in the ui but continue to receive
notifications.