Fixed a bug on single service override attribute.
When a single service_override_attribute was set on a host, it is
splitted on comma used as host/service delimiter, and attribute value
get false.

This patch adds an additionnal property attribute named `split_on_coma`
that indicates if a single value should be splitted on coma delimiter.

The `service_overrides` attribute is explicitely declared as non splitted.