add setup.py install --exclude-doc option
Add a new option --exclude-doc to install command of setup.py in order
to avoid installing the source files of the documentation (*.rst and
associated binary files).

Beware this is a workaround! The logic has been added in install class
since the documentation files must be considered as data files for
shinken cli and automatic build of documentation at module installation
time.

This option should be relatively helpful to distro packagers though.