Avoids collisions of Downtime.id between schedulers
Each scheduler starts Downtime.id from 1 and increments it by 1 at each
downtime creation. So, when running multiple schedulers, multiple
downtimes will have the same ID. Commands like DEL_SVC_DOWNTIME are
broadcasted to all schedulers for execution. When deleting a downtime,
if the ID is not unique, all the downtimes that have the same ID will be
deleted. By using microseconds since 1970-01-01 as an ID, the probability
of collision is reduced.