Fix: tests: guessed Notification id could be another value if others tests files were previously run in the same session/process.
If you run this test with others in the same process,
like with: "nosetests -xv test_no_host_template.py test_notification_master.py",
then Notification.id may be != 12 ! (why not 42 ? :p)

Just mock Notification.__init__ (the only one we are interested here) and retain the newly created instances in a list.
The one we'll need should be the first (guessed_notif).