Fix: (double-fix) sys.stdout has not necessarily always an 'encoding' attribute + shinken.log.obj must be checked before use.
1) Under nosetest:
sys.stdout can be replaced by a StringIO instance. Which lack an encoding attribute.

Then logger.error('Encoding detection error ..') was called.
But in shinken.log the 'obj' was still None -> None.add(..) -> attribute error.

Replaced the whole "try: except:" by a simple, good, single line :)

2) In shinken.log:
obj could still be uninitialized when a logger.{info/error/..} is called
-> None.add -> attribute error.
Check that "obj" is not "falsy".

OK not very important (because log has been rewritten in v2) but for the understanding of everybody.