Fix: don't error if servicedependency reference a service that is excluded/not included.
If a host specific service is either excluded (service_excludes) or not included (service_includes) in the host services,
then if this host is part of some servicedependency rule (as dependent or depended on) then that servicedependency must be disabled/dropped.

Otherwise you always get a "Service xxx not found for host yyy".