Enh: scheduler: don't create unnecessarily (many) copies of services and hosts when you want to iter over them
particularly this kind:

for elt in [y for y in [x for x in self.hosts] + [x for x in self.services]]:

which create a first list copy of hosts,
also another list copy of services,
then create a third one being the sum of the 2 previous ( [..] + [..] )
then create a fourth one [y for y in [..]+[..] ]
then iterate over the last one.

On big configs (more than thousands of services or hosts) this could actually quite help.