improve script directory finding in init scripts
the per-daemon init scripts delegate to the main `shinken` init
script. In order to do this, they find the current directory (where all
the init scripts are defined), and then calls `$curdir/shinken`.

The previous method of finding the current directory did not account for
symlinks, so running something like `/etc/rc2.d/S20shinken-poller
start` (where `S20shinken-poller` is a symlink)
would assume the current directory is `/etc/rc2.d/`, and then fail when
it couldn't find the main shinken init script.

This patch uses the method from http://stackoverflow.com/a/1638397 to
find the current directory of the script and resolve symlinks.

refs #1543