Fixed service_(includes|excludes) tmpl recursion
Fixed a regression introduced in
9f9bc2040289bd897a9064c6445ebc08221440f9

If an attrubute covered by a property is explicitely set at class or
instance level (during initialization for example), it breaks
attribute's template recursion in `get_property_by_inheritance` on
`if hasattr(self, prop)` (attributes is always seen as affected, even if
it was never set).

In the present case, it broke service exclusion defined on templates
rather than host.