Enh: implemented initial_state management
This patch adds management of the `initial_state` attribute on hosts and
services.

This attribute was available as an host or service attribute, but no
code behind took it into account.

An host or service initial state default value is still `PENDING`, but
it may be overridden for a specific host or service by setting its
`initial_state`. The initial output may also be configured via the
`initial_output` attribute.

Examples:

define host{
    use                          generic-host
    host_name                    test_host_0
    address                      127.0.0.1
    initial_state                u
    initial_output               Waiting for status
}

define service {
  use                            generic-service
  host_name                      test_host_0
  service_description            test_service_0
  initial_state                  o
  initial_output                 Waiting for status
}

**Caution**: the initial state of hosts or services are configured at
configuration parse level, but they may be overwritten by the scheduler
when the retention data is loaded if a previous state has been saved.