Push some more internal metrics into statsd
This patch adds (and extends) metrics sent to kernel.shinken.io into statsd,
such as number of elements in the  queues, number of managed objects, and so
on.

Those metrics are collected and sent to statsd by a dedicated thread, with
frequency controlled by the new `statsd_interval` in the global configuration
file (which defaults to 5s).