Memory leak hunding
This patch introduces several features to protect services from  memory leaks
in big environmensts, including queues throttling options, memory watchdog and
graceful restart.

**Queues throttling**

In big environments where pollers or reactionners manage many concurrent
actions, a synchronization effect may appear where a single poller or scheduler
gets all the scheduled actions at once. The result is a single machine drawing
almost all the workload, the other doing almost nothing.

This may lead to a chain reaction described below (let's illustrate it with a
poller, it's the same for reactionners).

- A poller gets more actions it can manage.
- Under high load, the concurrency creates a contention on the internal queues.
- When contention lowers (because actions get executed), big results
  lists are returned to the scheduler.
- The scheduler has to deserialize many objects at once to integrate them.
  As the python interpreter does not correctly free the used memory, this
  may generate a memory leak: object are deserialized into memory, processed,
  then deleted, but the used memory is not really freed by the interpreter.
- Each integrated result generates a new brok, which creates huge broks
  lists.
- The broker fetching all the broks at once, it has to deserialize huge broks
  lists, involving the same effect as on the scheduler.

The higher the load, the worst the effect.

In addition, if results are too long to come back, the actions get rescheduled
which aggravate the phenomenon.

This may lead to machines crash due to memory exhaustion.

To prevent this situation, throttling options have been added to control
the number of checks a poller or reactionner may get, the number of
results it may return, and the number of broks a broker may get. They
are described below.

*Poller and reactionner options*

- `max_q_size`: defines the maximum number of actions (slots) a poller (or
  reactionner) may hold.
- `q_factor`: same option as the previous one, but the number of
  available slots is calculated from the formula `workers *
  processes_by_worker * q_factor`, so it may automatically adapt to the
  number of CPU cores.
- `results_batch`: defines the maximum number of results a poller or
  reactionner may return at once.

*Broker options*

- `broks_batch`: defines the maximum number of broks a broker may get at once.

**Graceful restart**

Another memory leak comes when new configuration is sent from the arbiter.
As the configuration is made of many objects linked by circular references,
they are not properly garbage collected. To release this memory, the only
solution is to restart the service.

This patch introduces the graceful restart. If enabled, when a new
configuration is received by a daemon, it forks a new daemon and send it the
received configuration on stdin. Next it exits to let the newly spawned
instance replace it. This avoids having to wait for a new configuration
after a classical restart.

Graceful restart can be enabled by setting the `graceful_enabled` to `1` in
the daemon configuration.

**Memory watchdog**

To protect the machines from crashing because of memory exhaustion if the
previous options are not or badly set, a memory watchdog may be activated with
the `harakiri_threshold` in the services configuration. It automatically
restarts the services if their memory goes above the threshold.

**Actions priority**

The patch also introduces the checks, event handlers and notifications priority.
The commands may be added the `priority` attribute. When daemons ask the
scheduler for actions, it will send the actions with the lowest `priority`
first. This allows to guaranties that important actions will get executed
first.

**Note**: by default, the options above are not set, and the services
operate as usual.

**Note2**: this patch allows to manage slightly bigger configuration on
a single scheduler. It is no more necessary to spawn many scheduler to
spread the objects.