Fixed omission of min or max from graphs if set to 0
String version of Metric does not include min or max values, even
if they were present in the string passed to the constructor.

String version of Metric does not include values for warn, crit,
min, max if they were specified as 0. This happens because their
inclusion in the string representation of Metric was conditional
on their boolean value (which is false for 0). Changed to check
for None instead.

String version of Metric did not collapse None values into empty
semi-colon delimited sections, causing input of ;;1 to result
in ;1 on conversion to string.