Made memory free an opt-in option
As I'm not totally sure that freeing memory using low level libc6 calls
will work on all distributions, I prefer to let the user decide whether or
not to use it.

The call to `memory_free()` is now disabled by default, and may be
activated by setting the `aggressive_memory_management` option to `1` in
the daemons ini configuration file.

For instance, in `schedulerd.ini`:

    ...
    port=7768
    daemon_enabled=1
    aggressive_memory_management=1
    ...