Enh: Maintenance checks
This patch introduces maintenance checks, and extends the commands scheduling
with *check variants*: it allows to have commands executed by pollers,
but for other purpose than state check.

The *maintenance check* is an extension of the *maintenance period*, where
maintenance state is got from a command return code (as a of state check)
rather than a fixed predefined timeperiod. This allows to dynamically
get and host or service maintenance state from an arbitrary command. This
command can for instance query the state in an external database or API.

Example:

    define service {
        use                           generic-service
        host_name                     lb
        service_description           Web frontend
        check_command                 check_http_frontend
        maintenance_checks_enabled    1
        maintenance_check_command     check_release
        maintenance_check_interval    1
        maintenance_retry_interval    1
        maintenance_check_period      24x7
        ...
    }

    define command {
        command_name                  check_release
        command_line                  $PLUGINSDIR$/check_release_api
    }