Fix: http_client exceptions management
This patch fixes a bug in `http_client`: when an HTTP code != 200 is got,
an uncaught exception is risen, which results in a service crash.

The `Exception` instance is replaced by an `HTTPException`, which is
correctly handled. The `put()` method error management behavior has also
been harmonized with other methods.