Fix brok queues not producing broks
This commit patches a few operators in a few `if` conditionals that were
causing issues when broks_batch equals to 0.  The conditional test is
reversed, and that means that the test enters the wrong branch when
broks_batch equals to 0.

Because broks_batch equals to 0, `min(broks_batch, len(X))` will always
equal to 0 no matter how hard is tried.  As a consequence of this, the
broker will never be able to fetch broks from the scheduler, which means
that the broker will never have any valid data and will look
uninitialized.

Additionally, because the same conditional is reversed in the broker,
stuff such as external commands will never be successfully delivered to
the arbiter and other interested parties, because the broker will never
be able to provide more than 0 broks each time.

This may have been uncaught earlier if broks_batch is greater than 0
anyway, because then it at least returns something that can be consumed
by other agents part of Shinken.