Use kill -0 rather than ps to check process status
The `ps` command is not consistent across distributions: all versions do not
have the same parameters. For instance, `ps -p` is not supported under
BSD and alpine linux with its minimalist stdlib.

It's preferable to use the `kill` command with signal `0` to check process
status as it's universally the same across UNIX like systems.