Improved test docker files, setup.py, init script
Some of the docker test files could no longer work because some
dependencies could not be installed easily (those tests are supposed to
validate that the software can be installed without too much hassle on
various main distributions).

The remaining docker files needed some adaptation for the software to be
properly installed.

Also started to improve `setup.py` to make it more readable and
maintainable, and rely more on python standards when possible. The
shinken executables are now managed using the `scripts` parameter of the
`setup()` function. This way, the interpreter is the same as the one used for
the installation, which is far more reliable than trying to guess the
most resent <=2.7 python version.

The init script has also been updated to use more universal commands (`ps`
parameter, and `kill -0`) to make it more cross platform friendly. The
shinken executables are also looked for in the `PATH` environment
variable rather than looking for the `/usr/bin` directory. Coupled with
the use of the `scripts` function detailed above, this make it possible
to install and run shinken in virtualenvs.