Refactored the HTTP/serialization communications
Python3 is more sensible to data types, and the way POST queries were
made prevented the serialization/compression from working in binary
mode.

This patch refactores the way services communicate through HTTP, and
encapsulates all the factorization operations in a dedicated module.
This way, it'll be easier to change the serialization protocol in the
future.