WIP: Shinken python 3 port (#2021)
* Shinken python 3 port

This patch starts porting Shinken to pyhon3, with backward python2
compatibility

* Use kill -0 rather than ps to check process status

The `ps` command is not consistent across distributions: all versions do not
have the same parameters. For instance, `ps -p` is not supported under
BSD and alpine linux with its minimalist stdlib.

It's preferable to use the `kill` command with signal `0` to check process
status as it's universally the same across UNIX like systems.

The init script has also been updated to use more universal commands (`ps`
parameter, and `kill -0`) to make it more cross platform friendly. The
shinken executables are also looked for in the `PATH` environment
variable rather than looking for the `/usr/bin` directory. Coupled with
the use of the `scripts` function detailed above, this make it possible
to install and run shinken in virtualenvs.

* Workaround a bug with CherryPy < 9 and python3

A bug in CherryPy < 9.0.0 prevents using pyOpenSSLAdapter with Python3.

Here's the tail of the traceback got when importing this module:

    Traceback (most recent call last):
      ...
	from cherrypy.wsgiserver.ssl_pyopenssl import pyOpenSSLAdapter
      File "/usr/lib/python3/dist-packages/cherrypy/wsgiserver/ssl_pyopenssl.py", line 46, in <module>
	class SSL_fileobject(wsgiserver.CP_makefile):
    TypeError: function() argument 'code' must be code, not str

See https://github.com/cherrypy/cherrypy/issues/1399 for details.

This patch enforces the use of the `BuiltinSSLAdapter` with CherryPy < 9
under python3.

* Refactored the HTTP/serialization communications

Python3 is more sensible to data types, and the way POST queries were
made prevented the serialization/compression from working in binary
mode.

This patch refactores the way services communicate through HTTP, and
encapsulates all the factorization operations in a dedicated module.
This way, it'll be easier to change the serialization protocol in the
future.

* Added systemd unit management and optimized setup

- Used more standard paths for shinken additional files (modules, doc,
  examples, and so on...)
- Got rid of the optparse trick that was hiding the distutils command
  line parsing.
- Replaced the post-install steps from the install phase by
  a postinstall command that can be run independently.
- Factorized the configuration templates processing
- Got rid of the `--update` mode that was introducing the risk of
  overwriting the running configuration
- Isolated the distribution configuration in an examples directory

* Sandboxes virtualenvs

When installing from a virtualenv, install the files in the context of
the virtualenv, rather in the system.

As it's a more advanced use case, it's up to the admin to copy the
necessary files outside of the virtualenv.

* Made setup.py more os/dist aware

- Declined default paths per linux/bsd distribution
- Templated/parameterized init scripts/units to setup binaries absolute pathes
- Templated/parameterized configuration file paths
- Templated/parameterized default files paths
- Added the ability to install config files from examples
- Added the ability to install default files from examples
- Added the ability to install init scripts from examples
- Previous commands have no effect on Windows

* Updated installation documentation

Also fixed some obvious documentation syntax errors.

* Fixed init scripts deployement for sysv in setup.py